unfuck circular dep - move locateExtrudeDeclarator
